Create new module for modmail
