+++ praw/models/reddit/subreddit.py
@@ -5,13 +5,14 @@ import time
 
 from prawcore import Redirect
 
-from ...const import API_PATH, urljoin, urlparse
-from ...exceptions import APIException, ClientException
+from ...const import API_PATH, urljoin
+from ...exceptions import APIException
 from ..util import permissions_string, stream_generator
 from ..listing.generator import ListingGenerator
 from ..listing.mixins import SubredditListingMixin
 from .base import RedditBase
 from .mixins import MessageableMixin
+from .modmail import ModmailConversation
 from .wikipage import WikiPage
 
 
@@ -1533,85 +1534,6 @@ class Modmail(object):
 
     def __init__(self, subreddit):
         self.subreddit = subreddit
-
-
-class ModmailConversation(RedditBase):
-    """A class for modmail conversations."""
-    STR_FIELD = 'id'
-
-    @staticmethod
-    def id_from_url(url):
-        """Return the ID contained within a conversation URL.
-        :param url: A url to a conversation in the following format:
-            * https://mod.reddit.com/mail/all/2gmz
-        Raise :class:`.ClientException` if URL is not a valid conversation URL.
-        """
-        parsed = urlparse(url)
-        if not parsed.netloc:
-            raise ClientException('Invalid URL: {}'.format(url))
-
-        submission_id = parsed.path.rsplit('/', 1)[-1]
-
-        if not submission_id.isalnum():
-            raise ClientException('Invalid URL: {}'.format(url))
-        return submission_id
-
-    @classmethod
-    def parse(cls, data, reddit):
-        conversation = data['conversation']
-
-        conversation['authors'] = [reddit._objector.objectify(author)
-                                   for author in conversation['authors']]
-        conversation['owner'] = reddit._objector.objectify(
-            conversation['owner'])
-        conversation['user'] = reddit._objector.objectify(data['user'])
-        conversation.update(cls._convert_conversation_objects(data, reddit))
-
-        conversation = reddit._objector.snake_case_keys(conversation)
-
-        return cls(reddit, _data=conversation)
-
-    @classmethod
-    def _convert_conversation_objects(cls, data, reddit):
-        result = {'messages': [], 'modActions': []}
-        for object in data['conversation']['objIds']:
-            key = object['key']
-            object_data = data[key][object['id']]
-            result[key].append(reddit._objector.objectify(object_data))
-        return result
-
-    def __init__(self, reddit, id=None,  # pylint: disable=redefined-builtin
-                 url=None, _data=None):
-        super(ModmailConversation, self).__init__(reddit, _data)
-
-        if id is not None:
-            self.id = id  # pylint: disable=invalid-name
-        elif url is not None:
-            self.id = self.id_from_url(url)
-
-    def _fetch(self):
-        other = self._reddit.get(API_PATH['modmail_conversation']
-                                 .format(id=self.id))
-        self.__dict__.update(other.__dict__)
-        self._fetched = True
-
-
-class ModmailObject(RedditBase):
-    AUTHOR_ATTRIBUTE = 'author'
-    STR_FIELD = 'id'
-
-    def __setattr__(self, attribute, value):
-        if attribute == self.AUTHOR_ATTRIBUTE:
-            value = self._reddit._objector.objectify(value)
-        super(RedditBase, self).__setattr__(attribute, value)
-
-
-class ModmailAction(ModmailObject):
-    pass
-
-
-class ModmailMessage(ModmailObject):
-    pass
 
 
 class SubredditStream(object):
